mdx_gen/pyproject.toml
[project]
name = "mdx-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["F401"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
